requirements.txt
Pillow     9.2.0
pip        22.2.2
setuptools 49.2.1